[server][HatoholArmPluginGateHAPI2] Use correct type

+++ server/src/HatoholArmPluginGateHAPI2.cc
@@ -94,7 +94,7 @@ struct HatoholArmPluginGateHAPI2::Impl
 	const MonitoringServerInfo m_serverInfo;
 
 	ArmUtils m_utils;
-	ArmUtils::ArmTrigger m_armTrigger[static_cast<int>(HAPI2PluginCollectType::NUM_COLLECT_NG_KIND)];
+	ArmUtils::ArmTrigger m_armTrigger[static_cast<size_t>(HAPI2PluginCollectType::NUM_COLLECT_NG_KIND)];
 	HatoholArmPluginGateHAPI2 &m_hapi2;
 	ArmPluginInfo m_pluginInfo;
 	ArmFake m_armFake;
@@ -109,7 +109,7 @@ struct HatoholArmPluginGateHAPI2::Impl
 	Impl(const MonitoringServerInfo &_serverInfo,
 	     HatoholArmPluginGateHAPI2 &hapi2)
 	: m_serverInfo(_serverInfo),
-	  m_utils(_serverInfo, m_armTrigger, static_cast<int>(HAPI2PluginCollectType::NUM_COLLECT_NG_KIND)),
+	  m_utils(_serverInfo, m_armTrigger, static_cast<size_t>(HAPI2PluginCollectType::NUM_COLLECT_NG_KIND)),
 	  m_hapi2(hapi2),
 	  m_armFake(m_serverInfo),
 	  m_armStatus(),
